Move FAQ handlers to handlers package
